fix: address json boundary review comments
